Move mapper to its own file
